+++ package.json
@@ -68,7 +68,7 @@
     "draft-js": "^0.11.7",
     "draftjs-to-html": "^0.9.1",
     "express": "^4.17.1",
-    "firebase": "^8.0.0",
+    "firebase": "^7.20.0",
     "helmet": "^3.23.3",
     "html-to-draftjs": "^1.5.0",
     "moment": "^2.29.1",
